requirements.txt
fastapi>=0.104.1
pydantic>=2.5.2
python-dotenv>=1.0.0
uvicorn>=0.24.0
jinja2>=3.1.2
aiohttp>=3.9.1
openai>=1.3.7
rich>=13.7.0
typer>=0.9.0
python-multipart>=0.0.6
websockets>=12.0
pyyaml>=6.0.1
aiocron>=1.8
pandas>=2.1.4
scikit-learn>=1.3.2
torch>=2.1.2
transformers>=4.36.2
pillow>=10.1.0
numpy>=1.26.3
mangum>=0.17.0
